move import to top level and move event_loop
