+++ examples/add_shapes.py
@@ -7,10 +7,9 @@ your shapes.
 import numpy as np
 from skimage import data
 import napari
-from napari.util import gui_qt
 
 
-with gui_qt():
+with napari.gui_qt():
     # create the viewer and window
     viewer = napari.Viewer()
