+++ examples/add_vectors_image.py
@@ -7,11 +7,10 @@ Each vector position is defined by an (x-proj, y-proj) element
 """
 
 import napari
-from napari.util import gui_qt
-
 import numpy as np
 
-with gui_qt():
+
+with napari.gui_qt():
     # create the viewer and window
     viewer = napari.Viewer()
