+++ examples/change_image_dims.py
@@ -4,8 +4,7 @@ Change image shape and dims directly on the layer
 
 import numpy as np
 from skimage import data
-from napari import Viewer
-from napari.util import gui_qt
+from napari import Viewer, gui_qt
 
 
 with gui_qt():
