+++ examples/custom_key_bindings.py
@@ -4,10 +4,9 @@ Display one 4-D image layer using the add_image API
 
 from skimage import data
 import napari
-from napari.util import gui_qt
 
 
-with gui_qt():
+with napari.gui_qt():
     viewer = napari.Viewer()
     blobs = data.binary_blobs(
         length=128, blob_size_fraction=0.05, n_dim=2, volume_fraction=0.25
