+++ examples/dask_nD_image.py
@@ -6,10 +6,9 @@ import dask.array as da
 import numpy as np
 from skimage import data
 import napari
-from napari.util import gui_qt
 
 
-with gui_qt():
+with napari.gui_qt():
     blobs = da.stack(
         [
             data.binary_blobs(
