+++ examples/nD_image.py
@@ -5,10 +5,9 @@ Display one 4-D image layer using the add_image API
 import numpy as np
 from skimage import data
 import napari
-from napari.util import gui_qt
 
 
-with gui_qt():
+with napari.gui_qt():
     blobs = np.stack(
         [
             data.binary_blobs(
