+++ examples/nD_points.py
@@ -7,10 +7,9 @@ accross the dimensions, specified by their size
 import numpy as np
 from skimage import data
 import napari
-from napari.util import gui_qt
 
 
-with gui_qt():
+with napari.gui_qt():
     blobs = np.stack(
         [
             data.binary_blobs(
