+++ examples/nD_shapes.py
@@ -5,10 +5,9 @@ Display one 4-D image layer using the add_image API
 import numpy as np
 from skimage import data
 import napari
-from napari.util import gui_qt
 
 
-with gui_qt():
+with napari.gui_qt():
     blobs = data.binary_blobs(
         length=128, blob_size_fraction=0.05, n_dim=3, volume_fraction=0.1
     ).astype(float)
