+++ examples/nD_vectors.py
@@ -8,10 +8,9 @@ apprearing on each slice.
 import numpy as np
 from skimage import data
 import napari
-from napari.util import gui_qt
 
 
-with gui_qt():
+with napari.gui_qt():
     blobs = np.stack(
         [
             data.binary_blobs(
