+++ examples/set_theme.py
@@ -4,10 +4,9 @@ Displays an image and sets the theme to 'light'.
 
 from skimage import data
 import napari
-from napari.util import gui_qt
 
 
-with gui_qt():
+with napari.gui_qt():
     # create the viewer with an image
     viewer = napari.view(
         astronaut=data.astronaut(), multichannel=True, title='napari'
