+++ napari/__init__.py
@@ -9,6 +9,7 @@ with warnings.catch_warnings():
     from .viewer import Viewer
 
 from .view_function import view
+from ._qt import gui_qt
 
 from ._version import get_versions
