+++ napari/_qt/__init__.py
@@ -1,1 +1,2 @@
 from .qt_range_slider import QHRangeSlider, QVRangeSlider
+from .event_loop import gui_qt
