+++ napari/_qt/event_loop.py
