+++ napari/util/__init__.py
@@ -1,2 +1,1 @@
 from .misc import is_multichannel, segment_normal
-from .qt_event_loop import gui_qt
